Refactored ConfigurationFile unit tests - Still need to test on the camera
Now they will fail if not connected to a camera, or a CCF file is
missing, or a directory is missing.

The other class unit tests may take at least 2 weeks to refactor.

I also need to stop adding gratuitous context to my names, it is making
my code impossible to read.